[Silver II] Title: 에디터, Time: 104 ms, Memory: 4072 KB -BaekjoonHub

+++ 백준/Silver/1406. 에디터/에디터.cc
@@ -13,7 +13,7 @@ int main() {
 
 	stack<char> st;
 
-	for (size_t i = 0; i < s.size(); i++) {
+	for (int i = 0; i < s.size(); i++) {
 		st.push(s[i]);
 	}
